pyproject.toml
[tool.poetry]
name = "little-john"
version = "0.1.0"
description = ""
authors = ["will-ing <will-ing@github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pylint = "^2.5.3"
matplotlib = "^3.3.1"
keras = "^2.4.3"
tf-estimator-nightly = "^2.4.0-alpha.2020081801"
tf-nightly = "^2.4.0-alpha.20200818"
pytest-watch = "^4.2.0"
autopep8 = "^1.5.4"
finnhub-python = "^2.2.0"
mplfinance = "^0.12.7-alpha.0"
jupyterlab = "^2.2.5"
sklearn = "^0.0"
schedule = "^0.6.0"
figlet = "^0.0.1"
pyfiglet = "^0.8.post1"
termcolor = "^1.1.0"
pytest = "^6.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
